Help on decimal to binary function
def decimal_to_binary(number) means that the number will be provided by whomever uses the function.
In the function code that you write the number variable exists, and you can use it.  In this case the number will be a whole
number that will be converted to binary.  For this reason, you don't need line 21 num = 65, and change all num names
to number because num doesn't exist after you remove it from line 21.

the if statement is part of the for loop thus the code needs to be indented from line 26-30

There's no need for the print statement in line 33.