Help on is_prime function
In def is_prime(n): the n is a number that will be provided to the code when the function is used.
For this reason in line 83, 85, 88, and 90 number variable must be changed to n
Because the instructions state that LIne 92 happens after loop exits return True must have the same indentation as the for loop.
Why? It's not part of the for loop.